bob3

+++ src/main/java/org/example/vitalance/entidades/Alimento.java
@@ -21,9 +21,9 @@ public class Alimento {
 
     private String nombre; // en tu DER está como int, aunque lo lógico sería String
     private String descripcion;
-    private Double  carbohidrato;
-    private Long calorias;
-    private BigDecimal indiceGlucemico;
+    private Double carbohidrato;
+    private Double calorias;
+    private Double indiceGlucemico;
 
     //relacion con la tabla AlimentoComida,relacion bidireccional
     @OneToMany(mappedBy = "alimento",cascade = CascadeType.ALL,orphanRemoval = true)
